@@ -86,13 +86,13 @@
 		"email": "support@rocket.chat"
 	},
 	"devDependencies": {
-		"autoprefixer": "^7.2.3",
+		"autoprefixer": "^7.2.5",
 		"babel-mocha-es6-compiler": "^0.1.0",
 		"babel-plugin-array-includes": "^2.0.3",
-		"chimp": "^0.50.2",
+		"chimp": "^0.51.0",
 		"conventional-changelog-cli": "^1.3.5",
-		"eslint": "^4.13.1",
-		"mock-require": "^2.0.2",
+		"eslint": "^4.15.0",
+		"mock-require": "^3.0.1",
 		"postcss-custom-properties": "^6.2.0",
 		"postcss-import": "^11.0.0",
 		"postcss-media-minmax": "^3.0.0",
@@ -105,31 +105,74 @@
 		"supertest": "^3.0.0"
 	},
 	"dependencies": {
+		"@google-cloud/language": "0.8.0",
 		"@google-cloud/storage": "1.4.0",
-		"aws-sdk": "^2.172.0",
+		"@google-cloud/vision": "0.11.5",
+		"adm-zip": "0.4.7",
+		"atlassian-crowd": "0.5.0",
+		"autolinker": "1.6.0",
+		"aws-sdk": "^2.180.0",
 		"babel-runtime": "^6.26.0",
+		"bad-words": "1.3.1",
 		"bcrypt": "^1.0.3",
+		"bson": "0.5.5",
+		"bugsnag": "1.8.0",
 		"bunyan": "^1.8.12",
-		"codemirror": "^5.32.0",
-		"core-js": "2.5.3",
-		"emailreplyparser": "0.0.5",
+		"busboy": "0.2.14",
+		"cas": "https://github.com/kcbanner/node-cas/tarball/fcd27dad333223b3b75a048bce27973fb3ca0f62",
+		"clipboard": "1.7.1",
+		"codemirror": "^5.33.0",
+		"coffeescript": "1.12.7",
+		"connect": "2.7.10",
+		"core-js": "^2.5.3",
+		"csv-parse": "1.2.0",
+		"emailreplyparser": "^0.0.5",
 		"file-type": "^7.4.0",
+		"filesize": "3.5.11",
+		"gm": "1.23.1",
+		"gridfs-stream": "1.1.1",
+		"he": "1.1.1",
 		"highlight.js": "^9.12.0",
+		"iconv-lite": "0.4.19",
+		"image-size": "0.6.2",
 		"imap": "^0.8.19",
+		"ip-range-check": "0.0.2",
 		"jquery": "^3.2.1",
-		"ldapjs": "^1.0.1",
+		"jschardet": "1.6.0",
+		"ldapjs": "^1.0.2",
+		"less": "https://github.com/meteor/less.js/tarball/8130849eb3d7f0ecf0ca8d0af7c4207b0442e3f6",
+		"less-plugin-autoprefix": "1.5.1",
+		"localforage": "1.5.5",
+		"lokijs": "1.4.1",
+		"lru-cache": "2.6.5",
 		"mailparser": "^2.1.0",
+		"marked": "0.3.12",
 		"mime-db": "^1.32.0",
 		"mime-type": "^3.0.5",
+		"mkdirp": "0.5.1",
 		"moment": "^2.20.1",
 		"moment-timezone": "^0.5.14",
+		"node-dogstatsd": "0.0.7",
+		"object-path": "0.11.4",
 		"photoswipe": "^4.1.2",
 		"poplib": "^0.1.7",
 		"prom-client": "^10.2.2",
+		"querystring": "0.2.0",
+		"redis": "2.2.5",
 		"semver": "^5.4.1",
+		"slack-client": "2.0.6",
+		"speakeasy": "2.0.0",
+		"tar-stream": "1.5.5",
 		"toastr": "^2.1.4",
+		"twilio": "2.9.1",
+		"ua-parser-js": "0.7.17",
 		"underscore": "^1.8.3",
 		"underscore.string": "^3.3.4",
-		"wolfy87-eventemitter": "^5.2.4"
+		"wolfy87-eventemitter": "^5.2.4",
+		"xml-crypto": "0.6.0",
+		"xml2js": "0.2.0",
+		"xmlbuilder": "2.6.4",
+		"xmldom": "0.1.27",
+		"yaqrcode": "0.2.1"
 	}
 }
