@@ -16,7 +16,8 @@ Package.onUse(function (api) {
 
   api.use([
     'ddp',
-    'mongo'
+		'mongo',
+		'ecmascript'
   ], ['client', 'server']);
 
   api.use(['http', 'random'], 'web.cordova');
