@@ -27,8 +27,8 @@ Package.onUse(function(api) {
 
 });
 
+// It needs to be here cuz they are coffee files and need to be compiled
 Npm.depends({
-	'coffee-script': '1.10.0',
 	'hubot': '2.19.0',
-	'hubot-help': '0.2.0'
+	'hubot-help': '0.2.2'
 });
