@@ -1,5 +1,5 @@
 import _ from 'underscore';
-const dns = Npm.require('dns');
+import dns from 'dns';
 
 let emailDomainBlackList = [];
 let emailDomainWhiteList = [];
