@@ -1,20 +1,13 @@
 /*globals HTTPInternals, changeCase */
 import _ from 'underscore';
-
-const URL = Npm.require('url');
-
-const querystring = Npm.require('querystring');
+import URL from 'url';
+import querystring from 'querystring';
+import iconv from 'iconv-lite';
+import ipRangeCheck from 'ip-range-check';
+import he from 'he';
+import jschardet from 'jschardet';
 
 const request = HTTPInternals.NpmModules.request.module;
-
-const iconv = Npm.require('iconv-lite');
-
-const ipRangeCheck = Npm.require('ip-range-check');
-
-const he = Npm.require('he');
-
-const jschardet = Npm.require('jschardet');
-
 const OEmbed = {};
 
 //  Detect encoding
