@@ -34,10 +34,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', apiPublicChannelName);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', 0);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', apiPublicChannelName);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', 0);
 				channel._id = res.body.channel._id;
 			})
 			.end(done);
@@ -53,10 +53,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', apiPublicChannelName);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', 0);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', apiPublicChannelName);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', 0);
 			})
 			.end(done);
 	});
@@ -74,10 +74,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', apiPublicChannelName);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', apiPublicChannelName);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
@@ -155,10 +155,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', apiPublicChannelName);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', apiPublicChannelName);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
@@ -176,10 +176,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', apiPublicChannelName);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', apiPublicChannelName);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
@@ -210,7 +210,7 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('description', 'this is a description for a channel for api tests');
+				expect(res.body).to.have.nested.property('description', 'this is a description for a channel for api tests');
 			})
 			.end(done);
 	});
@@ -226,7 +226,7 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('topic', 'this is a topic of a channel for api tests');
+				expect(res.body).to.have.nested.property('topic', 'this is a topic of a channel for api tests');
 			})
 			.end(done);
 	});
@@ -242,7 +242,7 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('purpose', 'this is a purpose of a channel for api tests');
+				expect(res.body).to.have.nested.property('purpose', 'this is a purpose of a channel for api tests');
 			})
 			.end(done);
 	});
@@ -394,10 +394,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
@@ -428,9 +428,9 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
 			})
 			.end(done);
 	});
@@ -448,10 +448,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs);
 			})
 			.end(done);
 	});
@@ -469,10 +469,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs);
 			})
 			.end(done);
 	});
@@ -489,10 +489,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'c');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'c');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
@@ -510,10 +510,10 @@ describe('[Channels]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('channel._id');
-				expect(res.body).to.have.deep.property('channel.name', `EDITED${ apiPublicChannelName }`);
-				expect(res.body).to.have.deep.property('channel.t', 'p');
-				expect(res.body).to.have.deep.property('channel.msgs', roomInfo.channel.msgs + 1);
+				expect(res.body).to.have.nested.property('channel._id');
+				expect(res.body).to.have.nested.property('channel.name', `EDITED${ apiPublicChannelName }`);
+				expect(res.body).to.have.nested.property('channel.t', 'p');
+				expect(res.body).to.have.nested.property('channel.msgs', roomInfo.channel.msgs + 1);
 			})
 			.end(done);
 	});
