@@ -22,8 +22,8 @@ describe('[Direct Messages]', function() {
 			.expect(200)
 			.expect((res) => {
 				expect(res.body).to.have.property('success', true);
-				expect(res.body).to.have.deep.property('message.msg', 'This message was sent using the API');
-				expect(res.body).to.have.deep.property('message.rid');
+				expect(res.body).to.have.nested.property('message.msg', 'This message was sent using the API');
+				expect(res.body).to.have.nested.property('message.rid');
 				directMessage._id = res.body.message.rid;
 			})
 			.end(done);
